% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_anovat_lm.R
\name{lb_anovat_lm}
\alias{lb_anovat_lm}
\title{Create Custom ANOVA Table for Linear Regression}
\usage{
lb_anovat_lm(object, reg_collapse = TRUE, ...)
}
\arguments{
\item{object}{a Linear Regression object as returned from stats::lm()}

\item{reg_collapse}{Logical variable to collapse the elements in the table. Set to FALSE to see each variable and interaction in the model.}

\item{...}{other potential arguments}
}
\value{
a data frame with class anova that shows the Df, SS, MS, F and P for the regression
}
\description{
Create a standard ANOVA table for linear regression
}
\details{
This function is by Russell Steel for creating an ANOVA table for a full model linear regression.
}
\examples{
lb_anovat_lm(object = lm(mpg ~ disp * cyl, data = mtcars), reg_collapse = TRUE)
lb_anovat_lm(object = lm(mpg ~ disp * cyl, data = mtcars), reg_collapse = FALSE)
}
\references{
See https://community.rstudio.com/t/overall-anova-table-in-r/97896 and
https://community.rstudio.com/t/anova-table-for-full-linear-model/42074/3
}
