% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_create_lm_bands.R
\name{lb_create_lm_bands}
\alias{lb_create_lm_bands}
\title{Create data frame with confidence band data}
\usage{
lb_create_lm_bands(df, x, y, band_type = "whbands", level = 0.95)
}
\arguments{
\item{df}{a data frame containing at least three observations of two
variables for a linear regression}

\item{x}{a character name of an explanatory variable from the data frame}

\item{y}{a character name of a response variable from the data frame}

\item{band_type}{a character for the type of band desired. Includes:
"whbands".}

\item{level}{a number between 0 and 1 for the confidence level}
}
\value{
a data frame of the 100 x and predicted values along with lower
and upper values for the confidence band with level
}
\description{
This function takes input for a linear model and the type of confidence band
and alpha level and creates a data frame of values that can be plotted
to show the confidence bands.
}
\examples{
lb_create_lm_bands(mtcars, "disp", "mpg")
}
