% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_create_lm_bands.R
\name{lb_create_lm_bandsf}
\alias{lb_create_lm_bandsf}
\title{Create data frame with confidence bands data based on a linear model formula}
\usage{
lb_create_lm_bandsf(formula, data, band_type = "whbands", level = 0.95)
}
\arguments{
\item{formula}{a formula for a linear model with variables
from the data data frame}

\item{data}{a data frame containing at least three observations of two
variables for a linear regression}

\item{band_type}{a character for the type of band desired. Includes:
"whbands".}

\item{level}{a number between 0 and 1 for the confidence level}
}
\value{
a data frame of the 100 x and predicted values along with lower
and upper values for the confidence band with level
}
\description{
This function takes formula input for a linear model and the type of
confidence band and alpha level and creates a data frame of values that can
be plotted to show the confidence bands.
}
\examples{
lb_create_lm_bandsf(mpg ~ disp, data = mtcars)
}
