% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_plot_lm_bands.R
\name{lb_plot_lm_bands}
\alias{lb_plot_lm_bands}
\title{Plot linear model fitted values with confidence bands}
\usage{
lb_plot_lm_bands(df, x, y, band_type = "whbands", level = 0.95)
}
\arguments{
\item{df}{a data frame to be modeled}

\item{x}{a character name of an explanatory variable from the data frame}

\item{y}{a character name of a response variable from the data frame}

\item{band_type}{A character value for the type of confidence band.
Includes "whbands".}

\item{level}{a number between 0 and 1 for the confidence level of the bands}
}
\value{
a ggplot2 object with the fitted line, the confidence bands and
a loess smoother
}
\description{
Creates a plot object of a fitted linear model with confidence bands and
a loess smoother
}
\examples{
lb_plot_lm_bands(mtcars, "disp", "mpg")
}
