% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_plot_lm_bands.R
\name{lb_plot_lm_bandsf}
\alias{lb_plot_lm_bandsf}
\title{Plot linear model fitted values with confidence bands based on a formula}
\usage{
lb_plot_lm_bandsf(formula, data, band_type = "whbands", level = 0.95)
}
\arguments{
\item{formula}{a formula object to be fit by a linear model}

\item{data}{a data frame with the variables in the formula}

\item{band_type}{A character value for the type of confidence band.
Includes "whbands".}

\item{level}{a number between 0 and 1 for the confidence level of the bands}
}
\value{
a ggplot2 object with the points, the fitted line,
the confidence bands and a loess smoother
}
\description{
Creates a plot object of a fitted linear model with confidence bands and
a loess smoother based on a formula
}
\examples{
lb_plot_lm_bandsf(mpg ~ wt, data = mtcars)
}
