% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_conf_bands.R
\name{lb_whbands}
\alias{lb_whbands}
\title{Working-Hotelling bands for simple linear regression.}
\usage{
lb_whbands(object, new_data, level = 0.95)
}
\arguments{
\item{object}{An object of class "lm".}

\item{new_data}{A data frame containing the new data.}

\item{level}{The confidence level of the band.}
}
\value{
a list of 4 elements to include the predicted values as $fit
}
\description{
Intervals of the form "fit +/- w * standard-error", where w^2 is
found by \code{p * qf(level, p, n - p)}.
}
\examples{
lb_whbands(object = lm(mpg ~ disp, data = mtcars),
new_data = data.frame("disp" = seq(from = min(mtcars$disp),
to = max(mtcars$disp), length.out = 100)), level = .95)
}
\author{
David Gerard, American University
}
